企業APIを修正

+++ app/Services/CompanyService.php
@@ -3,44 +3,71 @@
 namespace App\Services;
 
 use App\Repositories\CompanyRepository;
+use App\Repositories\UserRepository;
 use Exception;
 use Illuminate\Support\Facades\Log;
 use Illuminate\Support\Facades\Validator;
 
-class CompanyService
+class CompanyService extends Service
 {
     /** @var \App\Repositories\CompanyRepository */
     private $companyRepository;
 
-    public function __construct(CompanyRepository $companyRepository)
-    {
+    /** @var \App\Repositories\UserRepository */
+    private $userRepository;
+
+    public function __construct(
+        CompanyRepository $companyRepository,
+        UserRepository $userRepository,
+    ) {
         $this->companyRepository = $companyRepository;
+        $this->userRepository = $userRepository;
     }
 
-    public function index(int $userId): \Illuminate\Http\JsonResponse
+    public function index(int $userId): \Illuminate\Database\Eloquent\Collection|array
     {
         try {
+            $user = $this->userRepository->find($userId);
+            if ( $user === null ) {
+                Log::error( __METHOD__ . ": User not found. (user_id={$userId})" );
+                return $this->errorUserNotFound();
+            }
+
             $companies = $this->companyRepository->getBy(['user_id' => $userId]);
 
-            return response()->ok($companies);
+            return $companies;
 
         } catch ( Exception $e ) {
             Log::error(__METHOD__);
             Log::error($e);
 
-            return response()->internalServerError();
+            return $this->errorInternalServerError();
         }
     }
 
-    public function show(int $userId, int $companyId): \Illuminate\Http\JsonResponse
+    public function show(int $userId, int $companyId): \App\Models\Company|array
     {
-        $company = $this->companyRepository->findBy(['user_id' => $userId, 'company_id' => $companyId]);
+        try {
+            $user = $this->userRepository->find($userId);
+            if ( $user === null ) {
+                Log::error( __METHOD__ . ": User not found. (user_id={$userId})" );
+                return $this->errorUserNotFound();
+            }
 
-        if ($company === null) {
-            return response()->notFound();
+            $company = $this->companyRepository->findBy(['user_id' => $userId, 'company_id' => $companyId]);
+            if ( $company === null ) {
+                Log::error( __METHOD__ . ": Company not found. (user_id={$userId}, company_id={$companyId})" );
+                return $this->errorCompanyNotFound();
+            }
+
+            return $company;
+
+        } catch ( Exception $e ) {
+            Log::error(__METHOD__);
+            Log::error($e);
+
+            return $this->errorInternalServerError();
         }
-
-        return response()->ok($company);
     }
 
     public function validateStore(array $postedParams): bool|array
@@ -60,26 +87,32 @@ class CompanyService
         $validator->setAttributeNames(['name' => '企業名']);
 
         if ( $validator->fails() ) {
-            return ['errors' => $validator->errors()->getMessages()];
+            return $this->errorBadRequest( $validator->errors()->getMessages() );
         }
 
         return true;
     }
 
-    public function store(int $userId, array $postedParams): \Illuminate\Http\JsonResponse
+    public function store(int $userId, array $postedParams): \App\Models\Company|array
     {
         try {
+            $user = $this->userRepository->find($userId);
+            if ( $user === null ) {
+                Log::error( __METHOD__ . ": User not found. (user_id={$userId})" );
+                return $this->errorUserNotFound();
+            }
+
             $params = array_merge(['user_id' => $userId], $postedParams);
 
             $company = $this->companyRepository->create($params);
 
-            return response()->ok($company->fresh());
+            return $company;
 
         } catch ( Exception $e ) {
             Log::error(__METHOD__);
             Log::error($e);
 
-            return response()->internalServerError();
+            return $this->errorInternalServerError();
         }
     }
 
@@ -100,59 +133,71 @@ class CompanyService
         $validator->setAttributeNames(['name' => '企業名']);
 
         if ( $validator->fails() ) {
-            return ['errors' => $validator->errors()->getMessages()];
+            return $this->errorBadRequest( $validator->errors()->getMessages() );
         }
 
         return true;
     }
 
-    public function update(int $userId, int $companyId, array $postedParams): \Illuminate\Http\JsonResponse
+    public function update(int $userId, int $companyId, array $postedParams): \App\Models\Company|array
     {
         try {
-            $company = $this->companyRepository->findBy(['user_id' => $userId, 'company_id' => $companyId]);
+            $user = $this->userRepository->find($userId);
+            if ( $user === null ) {
+                Log::error( __METHOD__ . ": User not found. (user_id={$userId})" );
+                return $this->errorUserNotFound();
+            }
 
+            $company = $this->companyRepository->findBy(['user_id' => $userId, 'company_id' => $companyId]);
             if ( $company === null ) {
-                return response()->notFound();
+                Log::error( __METHOD__ . ": Company not found. (user_id={$userId}, company_id={$companyId})" );
+                return $this->errorCompanyNotFound();
             }
 
             $isSuccess = $this->companyRepository->update($company, $postedParams);
 
             if ( ! $isSuccess ) {
-                throw new Exception( __METHOD__ . ": Failed update company. (company_id={$companyId})");
+                throw new Exception( __METHOD__ . ": Failed update company. (company_id={$companyId}, user_id={$userId}, posted_params=" . json_encode($postedParams, JSON_UNESCAPED_UNICODE) . ")");
             }
 
-            return response()->ok($company->fresh());
+            return $company->fresh();
 
         } catch ( Exception $e ) {
             Log::error(__METHOD__);
             Log::error($e);
 
-            return response()->internalServerError();
+            return $this->errorInternalServerError();
         }
     }
 
-    public function delete(int $userId, int $companyId): \Illuminate\Http\JsonResponse
+    public function delete(int $userId, int $companyId): \App\Models\Company|array
     {
         try {
-            $company = $this->companyRepository->findBy(['user_id' => $userId, 'company_id' => $companyId]);
+            $user = $this->userRepository->find($userId);
+            if ( $user === null ) {
+                Log::error( __METHOD__ . ": User not found. (user_id={$userId})" );
+                return $this->errorUserNotFound();
+            }
 
+            $company = $this->companyRepository->findBy(['user_id' => $userId, 'company_id' => $companyId]);
             if ( $company === null ) {
-                return response()->notFound();
+                Log::error( __METHOD__ . ": Company not found. (user_id={$userId}, company_id={$companyId})" );
+                return $this->errorCompanyNotFound();
             }
 
             $isSuccess = $this->companyRepository->delete($company);
 
             if ( ! $isSuccess ) {
-                throw new Exception( __METHOD__ . ": Failed delete company. (company_id={$companyId})");
+                throw new Exception( __METHOD__ . ": Failed delete company. (user_id={$userId}, company_id={$companyId})");
             }
 
-            return response()->ok($company);
+            return $company;
 
         } catch ( Exception $e ) {
             Log::error(__METHOD__);
             Log::error($e);
 
-            return response()->internalServerError();
+            return $this->errorInternalServerError();
         }
     }
 }
